requirements.txt
Pillow
sorl-thumbnail
